==> 1.data_collection.ipynb <==
# Data Collection

#### This cell is for defining various OPTIONS used for this notebook (working directory, how many rows and columns pandas displays for a dataframe, etc). 

#### Preferably this cell is also where we do important imports (for example pandas and numpy)

import os 
'''
Input the directory where your database is located 
'''
#os.chdir('C:/Users/Trond/Documents/Master 2020/Parkinson_PPMI')
os.chdir('C:/Users/Briggstone/Documents/Master 2020/Parkinson_PPMI')
#os.chdir('C:/Users/MyPC/Documents/Andrijana/UiS/DATMAS Master oppgave/Parkinson_PPMI')

'''
Where you want the csv file of the merged data to be placed
'''
#output_filepath = 'C:/Users/Trond/Documents/Master 2020/Processed data'
output_filepath = 'C:/Users/Briggstone/Documents/Master 2020/Processed data'
#output_filepath = 'C:/Users/MyPC/Documents/Andrijana/UiS/DATMAS Master oppgave/Processed data'

import pandas as pd 
import numpy as np

# Set ipython's max row display
pd.set_option('display.max_row', 1000)

# Set iPython's max column width to 150
pd.set_option('display.max_columns', 150)




#### In this cell we select the subset of patients with parkinson's that have valid enrollment dates

#Reading the data table Screening demographics into a pandas dataframe
data = pd.read_csv('_Subject_Characteristics/Screening___Demographics.csv') 

#Printing our the number of subjects total in the study
print("Number of total subjects in original data table: ", data.shape[0])

# Merging screening demographics wih enrollement data in order to select subjects with valid enrollment dates only
data2 = pd.read_csv('Study_Enrollment/Randomization_table.csv')
data = data.merge(data2, how = "inner", on = "PATNO")

#Discarding indviduals that do not have confirmed parkinson's, Everyone except APPRDX == 1 or APPRDX = 5. Also making sure that the subject has a valid enrollment date in the project
data = data.loc[((data.APPRDX == 1) | (data.APPRDX == 5))  & (data.ENROLLDT.notna())]

#Printing out number of subjects in the valid cohorts
print("Number of subjects with parkinson's in data table:", data.shape[0])
print("Number of subjects in De Novo Cohort:", data.loc[(data.APPRDX == 1) & (data.ENROLLDT.notna())].shape[0])
print("Number of subjects in Genetic Cohort:", data.loc[(data.APPRDX == 5) & (data.ENROLLDT.notna())].shape[0])

#Selecting only PATNO, APPRDX and ENROLLDT columns and reseting index (0-1-2 instead of 2-6-9).
pat_subset = pd.DataFrame(data.loc[:, ["PATNO", "ENROLLDT", "APPRDX"]].reset_index(drop = True))

#### In this cell we import our data tables into pandas DataFrames and select the features we want from each data table. 

#Importing data tables from non-motor and selecting relevant columns

# Montreal Cognitive Assesment
MOCA = pd.read_csv("Non-motor_Assessments/Montreal_Cognitive_Assessment__MoCA_.csv").loc[:,["PATNO", "EVENT_ID", "MCAALTTM", "MCACUBE", \
    "MCACLCKC", "MCACLCKN", "MCACLCKH", "MCALION", "MCARHINO", "MCACAMEL", "MCAFDS", "MCABDS", "MCAVIGIL", "MCASER7", "MCASNTNC", "MCAVF", \
    "MCAABSTR", "MCAREC1", "MCAREC2", "MCAREC3", "MCAREC4", "MCAREC5", "MCADATE", "MCAMONTH", "MCAYR", "MCADAY", "MCAPLACE", "MCACITY"]]

# Hopkins Verbal Learning Test
HVLT = pd.read_csv("Non-motor_Assessments/Hopkins_Verbal_Learning_Test.csv").loc[:,["PATNO", "EVENT_ID", "HVLTRT1", "HVLTRT2", \
        "HVLTRT3", "HVLTRDLY", "HVLTREC", "HVLTFPRL", "HVLTFPUN"]]

#Benton Judgement of Line
BJLO = pd.read_csv("Non-motor_Assessments/Benton_Judgment_of_Line_Orientation.csv").loc[:,["PATNO", "EVENT_ID", "JLO_TOTRAW"]]

#Letter Number Sequencing
LNSQ = pd.read_csv("Non-motor_Assessments/Letter_-_Number_Sequencing__PD_.csv").loc[:,["PATNO", "EVENT_ID", "LNS1A", "LNS1B", "LNS1C", \
        "LNS2A", "LNS2B", "LNS2C", "LNS3A", "LNS3B", "LNS3C", "LNS4A", "LNS4B", "LNS4C", "LNS5A", "LNS5B", "LNS5C", "LNS6A", "LNS6B", "LNS6C", \
        "LNS7A", "LNS7B", "LNS7C"]]

#Semantic Fluency
SEFL = pd.read_csv("Non-motor_Assessments/Semantic_Fluency.csv").loc[:,["PATNO", "EVENT_ID", "VLTANIM", "VLTVEG", "VLTFRUIT"]]

#REM Sleep Disorder Questionnaire
REMQ = pd.read_csv("Non-motor_Assessments/REM_Sleep_Disorder_Questionnaire.csv").loc[:,["PATNO", "EVENT_ID", "DRMVIVID", "DRMAGRAC", \
    "DRMNOCTB", "SLPLMBMV", "SLPINJUR", "DRMVERBL", "DRMFIGHT", "DRMUMV", "DRMOBJFL", "MVAWAKEN", "DRMREMEM", "SLPDSTRB", "STROKE", \
    "HETRA", "PARKISM", "RLS", "NARCLPSY", "DEPRS", "EPILEPSY", "BRNINFM", "CNSOTH"]]

# Geriatric Depression Scale Short
GDSS = pd.read_csv("Non-motor_Assessments/Geriatric_Depression_Scale__Short_.csv").loc[:,["PATNO", "EVENT_ID", "GDSSATIS", "GDSDROPD", \
    "GDSEMPTY", "GDSBORED", "GDSGSPIR", "GDSAFRAD", "GDSHAPPY", "GDSHLPLS", "GDSHOME", "GDSMEMRY", "GDSALIVE", "GDSWRTLS", "GDSENRGY", \
    "GDSHOPLS", "GDSBETER"]]

# University of Pennsylvania Smell Identification Test
SIDT = pd.read_csv("Non-motor_Assessments/University_of_Pennsylvania_Smell_ID_Test.csv").loc[:,["PATNO", "UPSITBK1", "UPSITBK2", \
    "UPSITBK3", "UPSITBK4"]]

# Epworth Sleepiness Scale
EPSS = pd.read_csv("Non-motor_Assessments/Epworth_Sleepiness_Scale.csv").loc[:,["PATNO", "EVENT_ID", "ESS1", "ESS2", \
    "ESS3", "ESS4", "ESS5", "ESS6", "ESS7", "ESS8"]]

# Scales for Outcomes in Parkinson’s Disease - Autonomic Dysfunction
SCOP = pd.read_csv("Non-motor_Assessments/SCOPA-AUT.csv").loc[:,["PATNO", "EVENT_ID", "SCAU1", "SCAU2", \
    "SCAU3", "SCAU4", "SCAU5", "SCAU6", "SCAU7", "SCAU8", "SCAU9", "SCAU10", "SCAU11", "SCAU12", "SCAU13", \
    "SCAU14", "SCAU15", "SCAU16", "SCAU17", "SCAU18", "SCAU19", "SCAU20", "SCAU21", "SCAU22", "SCAU23", "SCAU24", "SCAU25"]]

# Symbol Digit Modalities 
SDIM = pd.read_csv("Non-motor_Assessments/Symbol_Digit_Modalities.csv").loc[:,["PATNO", "EVENT_ID", "SDMTOTAL"]]


#Importing data tables from motor section

# MDS-Unified Parkinson's Disease Rating Scale
MSU3 = pd.read_csv("Motor___MDS-UPDRS/MDS_UPDRS_Part_III.csv").loc[:,["PATNO", "EVENT_ID", 'NP3BRADY', 'NP3FACXP', 'NP3FRZGT', \
    'NP3FTAPL', 'NP3FTAPR', 'NP3GAIT', 'NP3HMOVL', 'NP3HMOVR', 'NP3KTRML', 'NP3KTRMR', 'NP3LGAGL', 'NP3LGAGR', 'NP3POSTR', 'NP3PRSPL', 'NP3PRSPR', \
    'NP3PSTBL', 'NP3PTRML', 'NP3PTRMR', 'NP3RIGLL', 'NP3RIGLU', 'PN3RIGRL', 'NP3RIGN', 'NP3RIGRU', 'NP3RISNG', 'NP3RTALJ', 'NP3RTALL', 'NP3RTALU', \
    'NP3RTARL', 'NP3RTARU', 'NP3RTCON', 'NP3SPCH', 'NP3TTAPL', 'NP3TTAPR', 'PAG_NAME', 'PD_MED_USE', 'NHY']]
MSU1 = pd.read_csv("Motor___MDS-UPDRS/MDS_UPDRS_Part_I.csv").loc[:,["PATNO", "EVENT_ID", "NP1HALL", "NP1COG"]]
MSU2 = pd.read_csv ("Motor___MDS-UPDRS/MDS_UPDRS_Part_II__Patient_Questionnaire.csv").loc[:,["PATNO", "EVENT_ID", "NP2TRMR", "NP2WALK", "NP2FREZ"]]

#Importing data tables from Demographic data

# Family History
FMHS = pd.read_csv("_Subject_Characteristics/Family_History__PD_.csv").loc[:,["PATNO", "BIOMOMPD", "BIODADPD", \
    "FULSIBPD", "HAFSIBPD", "MAGPARPD", "PAGPARPD", "MATAUPD", "PATAUPD", "KIDSPD"]]

#Socio-Economics
SOEC = pd.read_csv("_Subject_Characteristics/Socio-Economics.csv").loc[:,["PATNO", "EDUCYRS"]]

# Screening Demographics
SCDE = pd.read_csv("_Subject_Characteristics/Screening___Demographics.csv").loc[:,["PATNO", "BIRTHDT", "GENDER"]]

# PD_Features
MDHI = pd.read_csv("Medical_History/PD_Features.csv").loc[:,["PATNO", "PDDXDT"]]

#### In this cell we do any additional preprocessing on the data before we perform our data merge

# SC (Screening) to BL (Baseline) for MOCA as MOCA is performed at screening, but we use as Baseline data
MOCA.loc[MOCA["EVENT_ID"] == "SC", "EVENT_ID"] = "BL"

#Drop MSU3 data entries that are post drug-administering

MSU3 = MSU3.drop(MSU3[ MSU3['PAG_NAME'] == "NUPDRS3A" ].index)
MSU3 = MSU3.drop("PAG_NAME", axis = 1)

#For Gender we need to combine 0 (Female of child bearing potential) and 1 (Female of non-child bearing potential) into one category 0 (female), we convert 2 (Male) to 1 afterwards
#Result of this processing: 0 = Female, 1 = Male

SCDE.loc[SCDE["GENDER"] == 1, "GENDER"] = 0 
SCDE.loc[SCDE["GENDER"] == 2, "GENDER"] = 1

#SCOP is not a simple sum and needs additional preprocessing to get total score
# In SCAU1-21, 9 is converted to 3. In SCAU22-25, 9 is converted to 0
for i in range(1,26):
    s = "SCAU" + str(i)
    points = 0
    if i < 22:
        points = 3
    SCOP.loc[SCOP[s] == 9, s] = points

#### In this cell we perform our data merge

# First step of join: Merge non-longitudinal data
data = pat_subset.merge(FMHS, how = "inner", on = "PATNO")
data = data.merge(SOEC, how = "inner", on = "PATNO")
data = data.merge(SCDE, how = "inner", on = "PATNO")
data = data.merge(SIDT, how = "inner", on = "PATNO")
data = data.merge(MDHI, how = "inner", on = "PATNO")

#Second step of join: Merge in longitudinal data

data = data.merge(MSU3, how = "inner", on ="PATNO")
data = data.merge(SDIM, how = "inner", on = ["PATNO", "EVENT_ID"])
data = data.merge(MSU1, how = "inner", on = ["PATNO", "EVENT_ID"])
data = data.merge(MOCA, how = "inner", on = ["PATNO", "EVENT_ID"])
data = data.merge(HVLT, how = "inner", on = ["PATNO", "EVENT_ID"])
data = data.merge(BJLO, how = "inner", on = ["PATNO", "EVENT_ID"])
data = data.merge(LNSQ, how = "inner", on = ["PATNO", "EVENT_ID"])
data = data.merge(SEFL, how = "inner", on = ["PATNO", "EVENT_ID"])
data = data.merge(GDSS, how = "inner", on = ["PATNO", "EVENT_ID"])
data = data.merge(EPSS, how = "inner", on = ["PATNO", "EVENT_ID"])
data = data.merge(SCOP, how = "inner", on = ["PATNO", "EVENT_ID"])
data = data.merge(REMQ, how = "inner", on = ["PATNO", "EVENT_ID"])
data = data.merge(MSU2, how = "inner", on = ["PATNO", "EVENT_ID"])



#### In this cell we do some additional processing post merge and output the merged data table to a csv file so that it can be further processed in missing_alues.ipynb

#Removing birthdate, date of Parkinson's disease diagnosis and project enrollment date and adding age and PD duration as a function of these columns. 

#Function for month/year accuracy of PD_DURATION
def pd_duration(row):
    
    enroll = float(row["ENROLLDT"][0:2]) / 12 + float(row["ENROLLDT"][-4:])
    pdstart = float(row["PDDXDT"][0:2]) / 12 + float(row["PDDXDT"][-4:])
    
    return enroll - pdstart
       
#Removing birthdate, date of Parkinson's disease diagnosis and project enrollment date and adding age and PD duration as a function of these columns. 

data ["AGE"] = data.apply(lambda row_wise: int(row_wise["ENROLLDT"][-4:]) - int(row_wise["BIRTHDT"]) , axis = 1)
data["PD_DURATION"] = data.apply(pd_duration, axis = 1)


#Updating AGE and PD_Duration for subjects by 1 year for each year in the study
year_dict = {
    "BL" : 0,
    "V04" : 1,
    "V06" : 2,
    "V08" : 3,
    "V10" : 4,
    "V12" : 5,
    "V13" : 6,
    "V14" : 7,
    "V15" : 8    
}

for k,v in year_dict.items():
    data.loc[data.EVENT_ID == k, "AGE"] = data.loc[data.EVENT_ID == k, "AGE"] + v
    data.loc[data.EVENT_ID == k, "PD_DURATION"] = data.loc[data.EVENT_ID == k, "PD_DURATION"] + v       
        
#Finally removing  birthdate, date of Parkinson's disease diagnosis and project enrollment date                                                       
data = data.drop(["ENROLLDT", "BIRTHDT", "PDDXDT"], axis = 1)




#Filtering out patients from genetic cohort that do not meet the inclusion criteria and then removing APPRDX as it no longer needed
valid_subjects = data.loc[(data.APPRDX == 1) | ((data.APPRDX == 5) & (data.AGE > 30) & (data.PD_DURATION <= 2) & (data.NHY < 3)  & (data.EVENT_ID == "BL") & \
                (data.PD_MED_USE == 0))].PATNO.unique()

#Series of Print statements that illustrate the result of filterting the genetic cohort
print("Subjects from Genetic Cohort with age over 30: ",data.loc[((data.APPRDX == 5) & (data.AGE > 30))].PATNO.unique().size)
print("Subjects from Genetic Cohort with age over 30 and PD duration less than 2: ",data.loc[((data.APPRDX == 5) & (data.AGE > 30) & (data.PD_DURATION <= 2))].PATNO.unique().size)
print("Subjects from Genetic Cohort with age over 30, PD duration less than 2 and nhy < 3: ", \
      data.loc[((data.APPRDX == 5) & (data.AGE > 30) & (data.PD_DURATION <= 2) & (data.NHY < 3))].PATNO.unique().size)
print("Valid subjects from Genetic Cohort: ",data.loc[((data.APPRDX == 5) & (data.AGE > 30) & (data.PD_DURATION <= 2) & (data.NHY < 3)  & (data.EVENT_ID == "BL") & \
                (data.PD_MED_USE == 0))].PATNO.unique().size)


#Filtering
for x in data.PATNO.unique():
    if x not in valid_subjects:
        data = data[data.PATNO != x]

#Removing APPDRX
data = data.drop("APPRDX", axis = 1)


#Dropping all PATNO who have hallucinations at baseline, NP1HALL != 0 or do not have baseline data

unique_subjects = data.PATNO.unique()

for u in unique_subjects:
    
    if data.loc[(data.PATNO == u) & (data.EVENT_ID == "BL"), :].shape[0] == 0:
        data = data[data.PATNO != u]
    elif data.loc[(data.PATNO == u) & (data.EVENT_ID == "BL") & (data.NP1HALL != 0), :].shape[0] != 0:
        data = data[data.PATNO != u]

        
'''
We are only interested in event_ids [BL,V04,V06,V08,V10,V12,V13,V14,V15] as these are the datapoints we can order in time and that has complete data
We therefore drop datapoints that do not have these event_ids
'''
valid_event_ids = ["BL", "V04", "V06", "V08", "V10", "V12", "V13", "V14", "V15"]
event_ids_to_remove = np.setdiff1d(data.EVENT_ID.unique(), valid_event_ids)

for x in event_ids_to_remove:
    data = data[data.EVENT_ID != x]

'''  
Adding a useful data column to the dataframe which tells us the last data point a subject has in the study
Useful for plotting study attrition for example
'''
data["LAST_EVENT_ID"] = np.nan
for x in data.PATNO.unique():
    subject_event_ids = data.loc[data.PATNO == x, "EVENT_ID"].unique()
    data.loc[data.PATNO == x, ["LAST_EVENT_ID"]] = (np.sort(subject_event_ids))[-1]


#We output the merged and partially processed data table to a file so that it can be processed further in missing_values.ipynb
data.to_csv(output_filepath + '/joined_data.csv', index = False)



### In this cell we can do any final data inspection (Number of unique patients, shape, etc)

import numpy as np
#Inspecting the merged data
print( "Data shape", data.shape)
print( "Number of unique PATNO:", data.PATNO.unique().size)

data.head(10)
